+++ src/libs/api/autorization/LoginApi.ts
@@ -14,7 +14,7 @@ export const pushData = () => {
   const data = new URLSearchParams()
   data.append('grant_type', 'authorization_code')
   data.append('client_id', VITE_CLIENT_ID)
-  data.append('redirect_uri', 'https://studay.me/')
+  data.append('redirect_uri', 'https://www.studay.me/login')
   data.append('code', getCode())
   data.append('client_secret', VITE_CLIENT_SECRET)
